Split context and imports
